inline: implement (2), fixed finalize_op(list)
forgot to push to @_ (gv rv2av args push)

change entersub to leave, leavesub to leave, always enclosing the inlined
sub into a block.  we can optionally skip that if certain ops are missing.

TODO: global *@_ needs to be localized and cleared, as it is used not in magic
padoffset 0.
use @_ not as ref