iter: make it a LOGOP
to please postfix re-ordering, which fixes up
the sibling and next ptrs.
This way peep and ck have less problems, even if it
wastes an unneded op_first ptr. This is true for all
LOGOPs btw. where op_first is only needed for kids during
compilation.

Note: This version is still a hack. We still have to fixup the loop
linklist (next ptrs) in rpeep.

Special-case newOP for OP_ITER in newWHILEOP to reserve 2 ptrs.
newUNOP would alloc an unneded STUB and newLOGOP is too heavy
and too early.

Also fix op_dump for iter.