Term-ReadKey: add cross probe for blockoptions
add a config.h, which can be used as temp. probe binary
for blockoptions(), if we have only a miniperl.
Note that for MSWin23 we hardcode the result, as the
compile/link step is too worrysome there.

See https://github.com/arsv/perl-cross/issues/40
and https://github.com/perl11/cperl/issues/214